gen: Move glib_wrapper() to glib_utils.rs
